requirements.txt
absl-py==0.12.0
appdirs==1.4.4
apturl==0.5.2
astroid==2.5.2
astunparse==1.6.3
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
evdev==1.3.0
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
idna==2.8
imageio==2.9.0
isort==5.8.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
key-mapper==0.8.1
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
netifaces==0.10.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas==1.2.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.15.7
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==1.9.73
pydbus==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.4.3
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
